Loading screen not updating - add error handling and remove fallbacks on baseUrl, version, apiKey (#228)
Debugging to see if the fallback values affect the project list not
being retrieved.
- Add STA_DYNAMO_API_URL to main.yml (to copy yesterdays changes made to
release.yml)
- Remove fallback logic when a variable is missing and instead throws an
error

Fixes a case when Dynamo is being used as the new production database,
however the request to fetch the data sometimes falls back to the
airtable when a variable is not defined in either .env or
app.config.json files - via https://www.the-sta.com/v1/projects

edit: some packages in this app are now too outdated to the point that
google no longer accepts builds to their store. For this reason, this PR
is being modified to upgrade many of these libraries so that
distribution to both stores is possible again.